+++ ZLYZhiHuDaily/ZLYZhiHuDaily/App/ZLYZhiHuDaily-Bridging-Header.h
@@ -10,6 +10,5 @@
 #define ZLYZhiHuDaily_Bridging_Header_h
 
 #import <SDWebImage/UIImageView+WebCache.h>
-//#import <MJRefresh/MJRefresh.h>
 
 #endif /* ZLYZhiHuDaily_Bridging_Header_h */
